+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log for Sidworks SearchResults plugin
 All notable changes to this module will be documented in this file.
 
+## 1.0.3 [22-05-2025]
+### Added
+- Results count has been added to keep track of the total results of a search term
+
 ## 1.0.2 [22-05-2025]
 ### Added
-- Enable/Disable ajax search logging. It's disabled by default.
+- Enable/Disable ajax search logging. It's disabled by default
 
 ### Changed
 - Refactored logger code